moved Search to its own component
