+++ src/components/Search/index.js
@@ -1,0 +1,6 @@
+// @flow
+import Search from './Search';
+
+export default {
+  Search,
+};
